EES-1815 Move releaseFile.ToFileInfo to ReleaseFileExtensions
