+++ src/GovUk.Education.ExploreEducationStatistics.Content.Model.Tests/Extensions/FileExtensionTests.cs
@@ -1,9 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using GovUk.Education.ExploreEducationStatistics.Common.Services;
 using Xunit;
-using static GovUk.Education.ExploreEducationStatistics.Common.Model.BlobInfo;
 using static GovUk.Education.ExploreEducationStatistics.Common.Model.FileType;
 using static GovUk.Education.ExploreEducationStatistics.Content.Model.Extensions.FileExtensions;
 
@@ -11,21 +9,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Tests.Extensi
 {
     public class FileExtensionTests
     {
-        private const string PublicationSlug = "publication-slug";
-        private const string ReleaseSlug = "release-slug";
-
-        private readonly ReleaseFile _releaseFile = new ReleaseFile
-        {
-            Release = new Release(),
-            File = new File
-            {
-                Id = Guid.NewGuid(),
-                RootPath = Guid.NewGuid(),
-                Filename = "ancillary.pdf",
-                Type = Ancillary
-            }
-        };
-
         [Fact]
         public void BatchesPath()
         {
@@ -186,67 +169,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Tests.Extensi
                     Assert.Throws<ArgumentOutOfRangeException>(() => file.PublicPath(release));
                 }
             });
-        }
-
-        [Fact]
-        public void ToFileInfo()
-        {
-            var result = _releaseFile.ToFileInfo(new BlobInfo
-            (
-                path: "Ignored",
-                size: "400 B",
-                contentType: "Ignored",
-                contentLength: -1L,
-                meta: new Dictionary<string, string>
-                {
-                    {NameKey, "Test ancillary file"}
-                }
-            ));
-
-            Assert.Equal(_releaseFile.FileId, result.Id);
-            Assert.Equal("pdf", result.Extension);
-            Assert.Equal("ancillary.pdf", result.FileName);
-            Assert.Equal("Test ancillary file", result.Name);
-            Assert.Equal("400 B", result.Size);
-            Assert.Equal(Ancillary, result.Type);
-        }
-
-        [Fact]
-        // TODO EES-1815 Remove this when the name is changed to use the database Subject name
-        public void ToFileInfo_BlobWithNoNameKey()
-        {
-            // Name is retrieved from the blob meta properties due to EES-1637 (Subject didn't exist early on and Name on Data files is Subject name persisted on blob)
-            // Chart files don't have any meta properties added to avoid duplicating data unnecessarily since nothing requires the filename or name of them
-            // Nevertheless test that we populate the name from the database File as a fallback when the blob name property is missing
-
-            var result = _releaseFile.ToFileInfo(new BlobInfo
-            (
-                path: "Ignored",
-                size: "400 B",
-                contentType: "Ignored",
-                contentLength: -1L,
-                meta: new Dictionary<string, string>()
-            ));
-
-            Assert.Equal(_releaseFile.FileId, result.Id);
-            Assert.Equal("pdf", result.Extension);
-            Assert.Equal("ancillary.pdf", result.FileName);
-            Assert.Equal("ancillary.pdf", result.Name);
-            Assert.Equal("400 B", result.Size);
-            Assert.Equal(Ancillary, result.Type);
-        }
-
-        [Fact]
-        public void ToFileInfoNotFound()
-        {
-            var result = _releaseFile.File.ToFileInfoNotFound();
-
-            Assert.Equal(_releaseFile.File.Id, result.Id);
-            Assert.Equal("pdf", result.Extension);
-            Assert.Equal("ancillary.pdf", result.FileName);
-            Assert.Equal("Unknown", result.Name);
-            Assert.Equal("0.00 B", result.Size);
-            Assert.Equal(Ancillary, result.Type);
         }
     }
 }
